brand-agent dynamic module
