--- packages/clarity-decode/types/brand-agent.d.ts
@@ -1,9 +1,0 @@
-import { PartialEvent } from "./core";
-
-export interface BrandAgentData {
-    name: string;
-    msg: string;
-    cid: string;
-}
-
-export interface BrandAgentEvent extends PartialEvent { data: BrandAgentData; }
