+++ packages/clarity-js/rollup.config.ts
@@ -23,7 +23,6 @@ export default [
       { file: pkg.module, format: "es", exports: "named" }
     ],
     plugins: [
-      alias({ entries: [ { find: /@src\/brand-agent.*/, replacement: '@src/brand-agent/blank' } ] }),
       resolve(),
       typescript(),
       commonjs({ include: ["node_modules/**"] })
@@ -41,7 +40,6 @@ export default [
       warn(message);
     },
     plugins: [
-      alias({ entries: [ { find: /@src\/brand-agent.*/, replacement: '@src/brand-agent/blank' } ] }),
       resolve(),
       typescript(),
       terser(terserOpts),
@@ -56,7 +54,6 @@ export default [
       warn(message);
     },
     plugins: [
-      alias({ entries: [ { find: /@src\/brand-agent.*/, replacement: '@src/brand-agent/blank' } ] }),
       resolve(),
       typescript(),
       terser(terserOpts),
@@ -78,8 +75,7 @@ export default [
           { find: '@src/layout/encode', replacement: '@src/insight/encode' },
           { find: /@src\/interaction\/(change|clipboard|input|pointer|selection)/, replacement: '@src/insight/blank' },
           { find: /@src\/layout.*/, replacement: '@src/insight/snapshot' },
-          { find: /@src\/performance.*/, replacement: '@src/insight/blank' },
-          { find: /@src\/brand-agent.*/, replacement: '@src/brand-agent/blank' }
+          { find: /@src\/performance.*/, replacement: '@src/insight/blank' }
         ]
       }),
       resolve(),
@@ -102,8 +98,7 @@ export default [
           { find: /@src\/layout.*/, replacement: '@src/performance/blank' },
           { find: /@src\/diagnostic.*/, replacement: '@src/performance/blank' },
           { find: /@src\/data\/(extract|baseline|summary)/, replacement: '@src/performance/blank' },
-          { find: /@src\/core\/dynamic/, replacement: '@src/performance/blank' },
-          { find: /@src\/brand-agent.*/, replacement: '@src/brand-agent/blank' }
+          { find: /@src\/core\/dynamic/, replacement: '@src/performance/blank' }
         ]
       }),
       resolve(),
@@ -173,7 +168,7 @@ export default [
     ]
   },
   {
-    input: "src/global.ts",
+    input: "src/dynamic/brand-agent/index.ts",
     output: [ { file: pkg["brand-agent"], format: "iife", exports: "named" } ],
     onwarn(message, warn) {
       if (message.code === 'CIRCULAR_DEPENDENCY') { return; }
