+++ packages/clarity-js/src/dynamic/brand-agent/index.ts
@@ -1,0 +1,4 @@
+import * as brandAgent from "@src/dynamic/brand-agent/brand-agent";
+(function () {
+  brandAgent.start();
+})();
