--- packages/clarity-js/types/brand-agent.d.ts
@@ -1,15 +1,0 @@
-// This should exactly match with Agent JS.
-
-export enum BrandAgentEventName {
-  BubbleShown,
-  NudgeClicked,
-  AgentMessageSent,
-  UserMessageSent,
-  ConversationStarted,
-}
-
-export interface BrandAgentData {
-  name: BrandAgentEventName;
-  msg?: string;
-  cid?: string;
-}
